backend/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.24
motor>=3.3
python-dotenv>=1.0
dnspython>=2.4
pydantic>=2.5
aiohttp>=3.9